backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.35
python-dotenv>=1.0.1
sse-starlette>=2.1.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
